Update timer.cpp
This addition allows for the user to know exactly what day the machine is working on to see if the timer is accurate. Also, the day of the week is an important aspect of the code as later code can use the specific day of the week to dispense particular increments of medicine.